# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a simple bookstore server with a book inventory, users and\
    \ a shopping cart."
  version: "1.0.0"
  title: "BookHub"
  contact:
    email: "info@bookhub.it"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
  /books:
    get:
      tags:
      - "book"
      summary: "Books filter"
      description: "Filter books by specified criteria"
      operationId: "booksGET"
      produces:
      - "application/json"
      parameters:
      - name: "title"
        in: "query"
        description: "Filter by name\n"
        required: false
        type: "string"
      - name: "not_in_stock"
        in: "query"
        description: "If true returns also books not in stock. Default is false."
        required: false
        type: "boolean"
      - name: "publishers"
        in: "query"
        description: "Filter by publishers' ID\n"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "authors"
        in: "query"
        description: "Filter by author"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "ISBN"
        in: "query"
        description: "Filter by ISBN"
        required: false
        type: "string"
      - name: "min_price"
        in: "query"
        description: "Filter by price higher than value"
        required: false
        type: "number"
      - name: "max_price"
        in: "query"
        description: "Filter by price lower than value"
        required: false
        type: "number"
      - name: "genre"
        in: "query"
        description: "Filter by genres"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "themes"
        in: "query"
        description: "Filter by themes"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "best_seller"
        in: "query"
        description: "If true returns bestsellers only"
        required: false
        type: "boolean"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        404:
          description: "No books found"
      x-swagger-router-controller: "Book"
  /books/{bookId}:
    get:
      tags:
      - "book"
      summary: "Find book by ID"
      description: "Returns a book"
      operationId: "getBookById"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "ID of book to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              book_id:
                type: "integer"
                format: "int64"
              isbn:
                type: "string"
              title:
                type: "string"
              price:
                type: "number"
                format: "double"
              price_currency:
                type: "string"
                description: "The currency of the price, in a string of length 3"
              picture:
                type: "string"
              abstract:
                type: "string"
              interview:
                type: "string"
              status:
                type: "string"
              publisher:
                type: "string"
              average_rating:
                type: "number"
                format: "double"
              authors:
                type: "array"
                description: "Array containing the full names of the authors of the\
                  \ book"
                items:
                  type: "string"
              genres:
                type: "array"
                description: "Array containing the genres of the book"
                items:
                  type: "string"
              themes:
                type: "array"
                description: "Array containing the themes of the book"
                items:
                  type: "string"
        400:
          description: "Invalid Request!"
        404:
          description: "Book not found"
      x-swagger-router-controller: "Book"
  /books/{bookId}/related:
    get:
      tags:
      - "book"
      summary: "Books similar to a specific book"
      description: "Collection of books similar to specified book"
      operationId: "relatedBooksGET"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "ID of the book similar to the ones returned"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of Books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        400:
          description: "Bad Input!"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Book"
  /books/{bookId}/events:
    get:
      tags:
      - "event"
      - "book"
      summary: "Get events of a book"
      description: "Given a book Id, returns all the events in which was presented"
      operationId: "bookEventsGET"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "Id of the book to get the events"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operations"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        400:
          description: "Invalid request"
        404:
          description: "No events found!"
      x-swagger-router-controller: "Event"
  /books/{bookId}/reviews:
    get:
      tags:
      - "review"
      - "book"
      summary: "Get reviews of a book"
      description: "Given a book Id, returns all the reviews"
      operationId: "bookReviewsGET"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "Id of the book to get the reviews"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        400:
          description: "Invalid request"
        404:
          description: "No reviews found!"
      x-swagger-router-controller: "Review"
...
  /user/add/address:
    post:
      tags:
      - "address"
      - "user"
      summary: "Add address to the current user"
      description: "Add an address to the user"
      operationId: "userAddAddressPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "addressStreetLine1"
        in: "formData"
        required: true
        type: "string"
      - name: "addressStreetLine2"
        in: "formData"
        required: false
        type: "string"
      - name: "city"
        in: "formData"
        required: true
        type: "string"
      - name: "zip_code"
        in: "formData"
        required: true
        type: "string"
      - name: "province"
        in: "formData"
        required: true
        type: "string"
      - name: "country"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful add"
        400:
          description: "Invalid request!"
        403:
          description: "Not authorized!"
      x-swagger-router-controller: "Address"
  /user/update/address:
    put:
      tags:
      - "address"
      - "user"
      summary: "Modify address of the current user"
      description: "Modify address of the current user.  If no parameters are specified\
        \ throws a 400 error"
      operationId: "userModifyAddressPUT"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "addressStreetLine1"
        in: "formData"
        required: false
        type: "string"
      - name: "addressStreetLine2"
        in: "formData"
        required: false
        type: "string"
      - name: "city"
        in: "formData"
        required: false
        type: "string"
      - name: "zip_code"
        in: "formData"
        required: false
        type: "string"
      - name: "province"
        in: "formData"
        required: false
        type: "string"
      - name: "country"
        in: "formData"
        required: false
        type: "string"
      responses:
        200:
          description: "Successful update!"
        400:
          description: "Invalid request!"
        403:
          description: "Not authorized!"
      x-swagger-router-controller: "Address"
...
  /cart:
    get:
      tags:
      - "cart"
      summary: "View the content of the cart"
      operationId: "cartGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "Max number of items to show"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Cart"
        400:
          description: "Invalid request!"
        403:
          description: "Unauthorized"
      x-swagger-router-controller: "Cart"
  /cart/add/book/{bookId}:
    put:
      tags:
      - "cart"
      summary: "Add/remove items to the cart"
      description: "Adds the given book by id to the cart. Optional quantity, defaults\
        \ to 1"
      operationId: "cartUpdatePUT"
      parameters:
      - name: "bookId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - name: "quantity"
        in: "query"
        description: "Number of copies to add to the cart. Default is 1."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Bad request"
        404:
          description: "Book not available in inventory"
      x-swagger-router-controller: "Cart"
  /cart/remove/book/{bookId}:
    delete:
      tags:
      - "cart"
      summary: "Remove items from the cart"
      description: "Remove the given book by id to the cart. Optional quantity, defaults\
        \ to 1"
      operationId: "cartRemoveDELETE"
      parameters:
      - name: "bookId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - name: "quantity"
        in: "query"
        description: "Number of copies to remove from the cart"
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Bad request"
        404:
          description: "Book not in the cart"
      x-swagger-router-controller: "Cart"
  /authors:
    get:
      tags:
      - "author"
      summary: "Get the list of authors"
      operationId: "authorsGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Author"
        404:
          description: "Author not found!"
      x-swagger-router-controller: "Author"
  /authors/{authorId}:
    get:
      tags:
      - "author"
      summary: "Get an author by ID"
      operationId: "authorIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "authorId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Author"
        400:
          description: "Invalid request!"
        404:
          description: "Author not found!"
      x-swagger-router-controller: "Author"
...
definitions:
...
  Cart:
    type: "object"
    required:
    - "books"
    - "total_amount"
    - "user_id"
    properties:
      user_id:
        type: "integer"
        format: "int64"
      total_amount:
        $ref: "#/definitions/Amount"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
    title: "Cart"
    description: "Order for books"
    example:
      books:
      - picture_path: "pic1.jpg"
        price:
          currency: "EUR"
          value: 11
        isbn: "123456679"
        genre:
          name: "Horror"
          description: "Very scary"
          genre_id: 1
        publisher:
          publisher_id: 1
          hq_location:
            country: "Italy"
            province: "CO"
            city: "Como"
            address_id: 1
            street_line2: "11"
            street_line1: "Via Valleggio"
            zip_code: "22100"
          name: "Zanichelli"
          description: "description"
        average_rating: 0.80082819046101150206595775671303272247314453125
        theme:
          name: "name"
          description: "description"
        book_id: 1
        title: "Il deserto dei tartari"
        authors:
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        status: "Available"
      - picture_path: "pic1.jpg"
        price:
          currency: "EUR"
          value: 11
        isbn: "123456679"
        genre:
          name: "Horror"
          description: "Very scary"
          genre_id: 1
        publisher:
          publisher_id: 1
          hq_location:
            country: "Italy"
            province: "CO"
            city: "Como"
            address_id: 1
            street_line2: "11"
            street_line1: "Via Valleggio"
            zip_code: "22100"
          name: "Zanichelli"
          description: "description"
        average_rating: 0.80082819046101150206595775671303272247314453125
        theme:
          name: "name"
          description: "description"
        book_id: 1
        title: "Il deserto dei tartari"
        authors:
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        status: "Available"
      user_id: 0
      total_amount:
        currency: "EUR"
        value: 11
  Book:
    type: "object"
    required:
    - "book_id"
    - "price"
    - "publisher"
    - "title"
    properties:
      book_id:
        type: "integer"
        format: "int64"
        example: 1
      isbn:
        type: "string"
        example: "123456679"
      title:
        type: "string"
        example: "Il deserto dei tartari"
      authors:
        type: "array"
        items:
          $ref: "#/definitions/Author"
      price:
        $ref: "#/definitions/Amount"
      picture_path:
        type: "string"
        example: "pic1.jpg"
        description: "Relative path to a picture representing the book\n"
      status:
        type: "string"
        example: "Available"
        description: "Book availability in the inventory"
        enum:
        - "Available"
        - "Out of stock"
      publisher:
        $ref: "#/definitions/Publisher"
      average_rating:
        type: "number"
        format: "double"
        description: "Average rating of the book, calculated from the reviews"
      genre:
        $ref: "#/definitions/Genre"
      theme:
        $ref: "#/definitions/Theme"
    title: "Book"
    description: "A book for sale in the store"
    example:
      picture_path: "pic1.jpg"
      price:
        currency: "EUR"
        value: 11
      isbn: "123456679"
      genre:
        name: "Horror"
        description: "Very scary"
        genre_id: 1
      publisher:
        publisher_id: 1
        hq_location:
          country: "Italy"
          province: "CO"
          city: "Como"
          address_id: 1
          street_line2: "11"
          street_line1: "Via Valleggio"
          zip_code: "22100"
        name: "Zanichelli"
        description: "description"
      average_rating: 0.80082819046101150206595775671303272247314453125
      theme:
        name: "name"
        description: "description"
      book_id: 1
      title: "Il deserto dei tartari"
      authors:
      - surname: "Rossi"
        name: "Mario"
        biography: "biography"
        picture: "picture"
      - surname: "Rossi"
        name: "Mario"
        biography: "biography"
        picture: "picture"
      status: "Available"
  Amount:
    type: "object"
    required:
    - "currency"
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000000000
      currency:
        $ref: "#/definitions/Currency"
    description: "Price of an item"
    example:
      currency: "EUR"
      value: 11
  Currency:
    type: "string"
    pattern: "^[A-Z]{3,3}$"
    description: "Currency international code\n"
    example: "EUR"
  Author:
    type: "object"
    required:
    - "name"
    - "surname"
    properties:
      name:
        type: "string"
        example: "Mario"
        description: "First name of the author\n"
      surname:
        type: "string"
        example: "Rossi"
        description: "Surname of the author\n"
      biography:
        type: "string"
        description: "Short biography of the author\n"
      picture:
        type: "string"
        description: "A portrait of the author\n"
    title: "Author"
    description: "An author that has written or co-written some books in the store"
    example:
      surname: "Rossi"
      name: "Mario"
      biography: "biography"
      picture: "picture"
  Address:
    type: "object"
    required:
    - "city"
    - "country"
    - "street_line1"
    - "zip_code"
    properties:
      address_id:
        type: "integer"
        format: "int64"
        example: 1
      street_line1:
        type: "string"
        example: "Via Valleggio"
      street_line2:
        type: "string"
        example: "11"
      city:
        type: "string"
        example: "Como"
      zip_code:
        type: "string"
        example: "22100"
      province:
        type: "string"
        example: "CO"
      country:
        type: "string"
        example: "Italy"
    title: "Address"
    description: "A complete address template, usable in many contexts of the web\
      \ app"
    example:
      country: "Italy"
      province: "CO"
      city: "Como"
      address_id: 1
      street_line2: "11"
      street_line1: "Via Valleggio"
      zip_code: "22100"
  Event:
    type: "object"
    required:
    - "date"
    - "event_id"
    - "location"
    - "name"
    - "presented_book"
    properties:
      event_id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Unique identifier of the event"
      location:
        $ref: "#/definitions/Address"
      presented_book:
        $ref: "#/definitions/Book"
      name:
        type: "string"
      date:
        type: "string"
        format: "date-time"
    title: "Event"
    description: "An event where some books are presented"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      event_id: 1
      name: "name"
      location:
        country: "Italy"
        province: "CO"
        city: "Como"
        address_id: 1
        street_line2: "11"
        street_line1: "Via Valleggio"
        zip_code: "22100"
      presented_book:
        picture_path: "pic1.jpg"
        price:
          currency: "EUR"
          value: 11
        isbn: "123456679"
        genre:
          name: "Horror"
          description: "Very scary"
          genre_id: 1
        publisher:
          publisher_id: 1
          hq_location:
            country: "Italy"
            province: "CO"
            city: "Como"
            address_id: 1
            street_line2: "11"
            street_line1: "Via Valleggio"
            zip_code: "22100"
          name: "Zanichelli"
          description: "description"
        average_rating: 0.80082819046101150206595775671303272247314453125
        theme:
          name: "name"
          description: "description"
        book_id: 1
        title: "Il deserto dei tartari"
        authors:
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        - surname: "Rossi"
          name: "Mario"
          biography: "biography"
          picture: "picture"
        status: "Available"
  Publisher:
    type: "object"
    required:
    - "name"
    - "publisher_id"
    properties:
      publisher_id:
        type: "integer"
        format: "int64"
        example: 1
      hq_location:
        $ref: "#/definitions/Address"
      name:
        type: "string"
        example: "Zanichelli"
        description: "Name of the editor/publisher"
      description:
        type: "string"
        description: "Some information on the publisher"
    title: "Publisher"
    description: "Who edits/publishes a book"
    example:
      publisher_id: 1
      hq_location:
        country: "Italy"
        province: "CO"
        city: "Como"
        address_id: 1
        street_line2: "11"
        street_line1: "Via Valleggio"
        zip_code: "22100"
      name: "Zanichelli"
      description: "description"
  Genre:
    type: "object"
    required:
    - "genre_id"
    - "name"
    properties:
      genre_id:
        type: "integer"
        format: "int64"
        example: 1
      name:
        type: "string"
        example: "Horror"
      description:
        type: "string"
        example: "Very scary"
    title: "Genre"
    description: "The genre of the book"
    example:
      name: "Horror"
      description: "Very scary"
      genre_id: 1
  Theme:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      description:
        type: "string"
    title: "Theme"
    description: "The theme of the book"
    example:
      name: "name"
      description: "description"
  Review:
    type: "object"
    required:
    - "book"
    - "rating"
    - "review_id"
    - "user"
    properties:
      review_id:
        type: "integer"
        format: "int64"
      user:
        type: "integer"
        format: "int64"
        description: "User id of who wrote the review"
      title:
        type: "string"
        description: "Title of the review"
      text:
        type: "string"
        description: "Content of the review"
      rating:
        type: "integer"
        format: "int64"
        description: "Rating of the review, expressed in 1 to 5 stars"
        minimum: 1
        maximum: 5
      book:
        type: "integer"
        format: "int64"
        description: "Id of the reviewed book"
    title: "Review"
    description: "A review of a book, with rating and optional title and text"
    example:
      review_id: 0
      book: 5
      rating: 1
      text: "text"
      title: "title"
      user: 6
...